feat: handle assets with multiple Fairplay encryption keys
Content owners mandate the use of separate encryption keys for SD and HD (or UHD) video tracks, so that different encryption policies can be enforced (eg. HDCP level).
This feature allows that by extracting the KeyID value from the skd:// URL and passing that as a query parameter in the license request.